pyproject.toml
[tool.poetry]
name = "unitest"
version = "0.1.0"
description = ""
authors = ["jggomez <juan.gomez01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-firestore = "^2.20.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
